Swap old page-build for new menu build
